Add documentation to all public items
